feat(db): enable with-statement for cleaner connection handling

+++ main.py
@@ -322,7 +322,7 @@ class SlotStore:
             self.conn.executescript(
                 """
                 CREATE TABLE IF NOT EXISTS users (
-                    id INTEGER PRIMARY KEY SERIAL,
+                    id INTEGER PRIMARY KEY BIGSERIAL,
                     username TEXT NOT NULL UNIQUE,
                     password_salt TEXT NOT NULL,
                     password_hash TEXT NOT NULL,
@@ -361,7 +361,7 @@ class SlotStore:
                 );
 
                 CREATE TABLE IF NOT EXISTS spin_results (
-                    id INTEGER PRIMARY KEY SERIAL,
+                    id INTEGER PRIMARY KEY BIGSERIAL,
                     user_id INTEGER NOT NULL,
                     difficulty_mode TEXT NOT NULL,
                     win_amount REAL NOT NULL,
